CF1-102 Fix passing tests using end methods
Fixes tests that use the `end` method of supertest. This method wasn't
working as expected, and it didn't seem to respect its preceding
`expect` method.

With these changes, the methods that used supertest's `end` method have
been changed to promises, by adding the `done` function parameter to the
missing tests, and by asserting what the `end` blocks were asserting in
a `then` block. That change makes these assertions synchronous, and will
fail if the promise rejects.